requirements.txt
livekit-agents==0.*
livekit-plugins-openai==0.*
livekit-plugins-silero==0.*
livekit-api==0.*
python-dotenv
livekit-api
flask[async]
flask
flask-cors
uvicorn
openai>=1.12.0